Refactored all source files into a src folder for better organization
