+++ index.js
@@ -1,6 +1,6 @@
 /* eslint-disable global-require */
 
 module.exports = {
-  replayProfile: require('./dist/replay').default,
-  interceptFetchCalls: require('./dist/intercept').default,
+  replayProfile: require('./dist/src/replay').default,
+  interceptFetchCalls: require('./dist/src/intercept').default,
 };
