+++ src/intercept.js
