+++ src/stringSimilarity.js
